moved api request to App.jsx, people list now doesn't refresh navigating to dashboard
